projects/quantsys/services/mcp_bus/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-jose>=3.3.0
PyJWT>=2.8.0
redis>=5.0.0
psutil>=5.9.0
langsmith>=0.6.4
